[Bronze V] Title: 이상한 기호, Time: 0 ms, Memory: 1112 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.cc
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main() {
-    int a, b, c, d;
-    scanf("%d %d", &a, &b);
-    printf("%d", (a+b) * (a-b));
+    long long int a, b, c, d;
+    scanf("%lld %lld", &a, &b);
+    printf("%lld", (a+b) * (a-b));
 }
